+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Allow destination address to come from a command.
+- Don't retry connections for non-temporary address errors (missing port, etc).
 
 # v0.11 (2019-02-19)